requirements.txt
requests[security]
retrying
PyJWT==2.3.0
cryptography==35.0.0
# Dev requirements
dash
pytest
pytest-mock
